Rename graphDataStruct to graphSchema
